Re-organize transformation files
