+++ src/transformation/transform/transform_stream_cipher.rs
